add eslint+prettier option, change template names
