+++ templates/default/.gitignore
@@ -1,9 +1,8 @@
 .idea
 .git
-.DS_Store
-.history
-.firebase
 build
-dist
 node_modules
+.DS_Store
+dist/
+.history
 *.log
