--- templates/default/.prettierrc.json
@@ -1,18 +1,0 @@
-{
-	"arrowParens": "avoid",
-	"bracketSpacing": true,
-	"htmlWhitespaceSensitivity": "css",
-	"insertPragma": false,
-	"jsxBracketSameLine": false,
-	"jsxSingleQuote": false,
-	"printWidth": 180,
-	"proseWrap": "preserve",
-	"quoteProps": "as-needed",
-	"requirePragma": false,
-	"semi": true,
-	"singleQuote": true,
-	"tabWidth": 2,
-	"trailingComma": "none",
-	"useTabs": true,
-	"resolveGlobalModules": true
-}
