+++ templates/lit-app/.vscode/extensions.json
