+++ templates/lit-app/.vscode/settings.json
