+++ templates/lit-app/public/manifest.json
