+++ templates/lit-app/public/service-worker.js
