+++ templates/lit-app/res/drawable/logo-192x192.png
