+++ templates/lit-app/res/drawable/logo-48x48.png
