+++ templates/lit-app/res/drawable/logo-512x512.png
