+++ templates/lit-app/res/drawable/og/default.png
