+++ templates/lit-app/src/base/app-404.ts
