+++ templates/lit-app/src/base/styles/default-theme.ts
