+++ templates/lit-app/src/base/styles/lit-shared-styles.ts
