+++ templates/lit-app/src/xconfig/strings/index.ts
