+++ templates/lit-app/src/xconfig/strings/lang/es.ts
