+++ templates/lit-app/src/xconfig/typings.d.ts
