--- templates/rapid-prototype/README.md
@@ -1,3 +1,0 @@
-# {{name}}
-
-{{description}}
