--- templates/rapid-prototype/index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>{{capital name space=true}} Lit Element Simple Starter Kit</title>
-</head>
-<body>
-    <{{name}}></{{name}}>
-    <script type="module" src="/src/index.ts"></script>
-</body>
-</html>
