--- templates/rapid-prototype/vite.config.ts
@@ -1,3 +1,0 @@
-import { defineConfig } from "vite";
-
-export default defineConfig({});
